refactored model to class
