+++ cnn_skin_cancer/src/model/utils.py
@@ -1,9 +1,11 @@
 from enum import Enum
 
-from basic_model_sub import BasicNet
-from crossval_model import *
 from keras.models import Model
-from resnet50_model_sub import ResNet50
+from model.basic_model import BasicNet
+from model.resnet50_model import ResNet50
+
+# https://towardsdatascience.com/tensorflow-class-inheritance-beautiful-code-59d2eb7cdfce
+# https://towardsdatascience.com/model-sub-classing-and-custom-training-loop-from-scratch-in-tensorflow-2-cc1d4f10fb4e
 
 
 class Model_Class(Enum):
@@ -26,6 +28,7 @@ def get_model(model_name: str, model_params: dict) -> Model:
                 loss=model_params.get("loss"),
                 metrics=model_params.get("metrics"),
             )
+            # TODO: doesnt need to print every time
             print(model.build_graph(model_params).summary())
             return model
